Fix: 'thousand item' bug
On non-whole unit items, a number >= 1000 would not appear on form but would appear on report. The cell was set to 'number' input, but my formatting would add a comma, so the returned string would not be valid for that input. Fixed issue by setting any item with a non-whole unit to be of type 'text' rather than of type 'number'. Does not change the way the number is stored in data so should not affect anything else. Could possibly push through a 'NaN' but this would not be catastrophic just confusing error to user. Also adjusted cell width.